Re-package v6.0.0 to include runtime files only.
The ExtJS 6.0.0 upstream release made some farily substantial changes to the
way the ZIP file was arranged, but the webjar pom.xml had not been updated
accordingly, and was creating a 200Mb artifact as a result (including all
examples, etc.).

This change packages up only the build/ directory (minus examples) along with
the bootstrap and license files from the root.